move schedule to schedule.rs
